fix(tombos-controller): adjust the relation of identifiers
